source list view allows file drop
